requirements.txt
aniso8601==9.0.1
Babel==2.14.0
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-BabelEx==0.9.4
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security==3.0.0
Flask-Security-Too==5.3.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
passlib==1.7.4
pytz==2023.3.post1
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.23
typing_extensions==4.9.0
Werkzeug==3.0.1
WTForms==3.1.1